pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pgbenchmark"
version = "0.0.4"
description = "A Python package to benchmark query performance and comparison on PostgreSQL Database"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache License 2.0"}
authors = [
    {name = "Elguja Lomsadze", email = "lomsadze.guja@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Database :: Front-Ends",
    "Topic :: Software Development :: Testing"
]
dependencies = [
    "psycopg2-binary>=2.9.10",
    "SQLAlchemy>=2.0.40",
    "click>=8.1.3",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0"
]

[project.urls]
"Homepage" = "https://github.com/GujaLomsadze/pgbenchmark"

[project.scripts]
pgbenchmark = "pgbenchmark.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
pgbenchmark = ["visualizer/static/*.html", "visualizer/static/css/*.css", "visualizer/static/js/*.js"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"